3654 - separate concerns for History (Hooks, Items)
