+++ src/client/components/Navigation/History/History.tsx
@@ -1,35 +1,12 @@
-// Placeholder component moved to History.tsx
-// This is still a placeholder component.
-
-import React, { useEffect } from 'react'
+import React from 'react'
 import { useTranslation } from 'react-i18next'
 
-import { ApiEndPoint } from 'meta/api/endpoint'
-import { ActivityLog, ActivityLogMessage } from 'meta/assessment'
-
-import { useAppDispatch } from 'client/store'
 import { useHistory } from 'client/store/data/hooks/useHistory'
-import { TablePaginatedActions, useTablePaginatedData } from 'client/store/ui/tablePaginated'
-import { useCountryRouteParams, useSectionRouteParams } from 'client/hooks/useRouteParams'
+import Items from 'client/components/Navigation/History/Items'
 import { useMaxHeight } from 'client/components/Navigation/NavAssessment/hooks/useMaxHeight'
-import RecentActivityItem from 'client/pages/CountryHome/FraHome/RecentActivity/RecentActivityItem'
 
-const path = ApiEndPoint.CycleData.activities()
-const useGetData = () => {
-  const dispatch = useAppDispatch()
-  const { assessmentName, cycleName, countryIso } = useCountryRouteParams()
-
-  useEffect(() => {
-    const params = { assessmentName, cycleName, countryIso, path, page: 0, limit: 20 }
-    dispatch(TablePaginatedActions.getData(params))
-  }, [assessmentName, countryIso, cycleName, dispatch])
-}
-
-const useData = () => {
-  const { sectionName } = useSectionRouteParams()
-  const data = useTablePaginatedData<ActivityLog<never>>(path)
-  return data?.filter((d) => d.message === ActivityLogMessage.descriptionUpdate && d.section === sectionName)
-}
+import { useData } from './hooks/useData'
+import { useGetData } from './hooks/useGetData'
 
 const History: React.FC = () => {
   const { t } = useTranslation()
@@ -48,13 +25,7 @@ const History: React.FC = () => {
             return <div key={h.sectionKey}> {t(h.sectionLabelKey)} </div>
           })}
         </div>
-        <div>
-          {data?.map((d, i) => (
-            <div key={d.time} className="nav-section__item">
-              <RecentActivityItem datum={d} rowIndex={i} />
-            </div>
-          ))}
-        </div>
+        <Items values={data} />
       </div>
     </div>
   )
