+++ src/client/components/Navigation/History/Items/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Items'
