+++ src/client/components/Navigation/History/hooks/useData.ts
@@ -1,0 +1,13 @@
+import { ApiEndPoint } from 'meta/api/endpoint'
+import { ActivityLog, ActivityLogMessage } from 'meta/assessment'
+
+import { useTablePaginatedData } from 'client/store/ui/tablePaginated'
+import { useSectionRouteParams } from 'client/hooks/useRouteParams'
+
+const path = ApiEndPoint.CycleData.activities()
+
+export const useData = () => {
+  const { sectionName } = useSectionRouteParams()
+  const data = useTablePaginatedData<ActivityLog<never>>(path)
+  return data?.filter((d) => d.message === ActivityLogMessage.descriptionUpdate && d.section === sectionName)
+}
